djangosurveybackend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dependency-injector==4.41.0
Django==3.2.23
django-cors-headers==4.1.0
django-health-check==3.16.5
django-ipware==5.0.0
django-log-viewer==1.1.7
django-redis==5.4.0
django-rest-swagger==2.2.0
django-structlog==5.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
drfpasswordless==1.5.9
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
pkgutil_resolve_name==1.3.10
psycopg2-binary
pycodestyle==2.9.1
pyflakes==2.5.0

pyrsistent==0.19.3
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
structlog==23.1.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
django-apscheduler==0.6.2
APScheduler~=3.10.4
django-jazzmin==2.6.0
pytest
isort
black
flake8
pytest-cov